battery monitor: move voltage-divider function to separate file
This prepares the battery monitor library to support other sources of
voltage measurement.
